awscdk-minecraft/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=46.1.0", "wheel", "build", "docutils"]

[project]
name = "awscdk-minecraft"
version = "2.2.0"
readme = "README.md"
keywords = ["Minecraft", "AWS CDK"]
classifiers = [
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10"
]
dependencies = [
    "importlib-metadata; python_version<'3.8'",
    "aws-cdk-lib >=2.45.0, <3.0.0",
    "constructs >=10.0.5, <11.0.0",
    "aws_cdk.aws_lambda_python_alpha",
    "aws_prototyping_sdk.static_website",
    "cdk-ecr-deployment >= 2.0.0"
]

[project.optional-dependencies]
test = ["pytest", "pytest-cov", "pytest-xdist"]
publish = ["twine"]
dev = ["pytest", "pytest-cov", "pytest-xdist", "twine"]
all = ["pytest", "pytest-cov", "pytest-xdist", "twine"]